Update destination.py
This hotfix addresses a bug I encountered when trying to load assessments into SC using a Snowflake-sourced ID-xwalk. 

```
TypeError: Cannot setitem on a Categorical with a new category (), set the categories first
```

For some reason, a Categorical column in being introduced during processing, and must be converted to a string before filling NAs.