Make YAML parser use config file path to render Jinja
If Jinja is imported in an earthmover config YAML, installing that package via
`earthmover deps` will break, because the relative imports cannot be found from
the root directory of the importing project (e.g.  student_id_wrapper).

This commit modifies `JinjaEnvironmentLoader.load_config_file` to pass the
parent directory of the config file being rendered to
`util.build_jinja_template`, which has also been modified to pass this
parameter to `jinja2.FileSystemLoader`. This replaces the original behavior,
where `FileSystemLoader` was always initialized with the current working
directory of the earthmover invocation ('./').